requirements.txt
anyio==3.5.0
asgiref==3.5.0
autopep8==1.6.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
fastapi==0.75.0
flake8==4.0.1
h11==0.13.0
httptools==0.4.0
idna==3.3
mccabe==0.6.1
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
python-dotenv==0.19.2
PyYAML==6.0
requests==2.27.1
sniffio==1.2.0
soupsieve==2.3.1
starlette==0.17.1
toml==0.10.2
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8
websockets==10.2
